user milestone/streak endpoint, add day on calender and include num of bottles taken, moved content of /relapse to progress
