+++ backend/app/api/router.py
@@ -9,7 +9,7 @@ from api.home.routers import router as home_router
 from api.example.views import router as example_router
 from api.progress.views import router as progress_router
 from api.call.views import router as call_router
-from api.relapse.views import router as relapse_router
+# from api.relapse.views import router as relapse_router
 
 from fastapi.security import OAuth2PasswordRequestForm
 from fastapi import Depends, FastAPI, HTTPException, status
@@ -47,4 +47,4 @@ api_router.include_router(progress_router,
                           prefix="/progress",
                           tags=["Progress"])
 api_router.include_router(contact_router, prefix="/contact", tags=["Contact"])
-#api_router.include_router(relapse_router, prefix="/relapse", tags=["relapse"])
+# api_router.include_router(relapse_router, prefix="/relapse", tags=["relapse"])
